# Conflicts:
#	pulsar-broker-common/src/main/java/org/apache/pulsar/broker/ServiceConfiguration.java
#	pulsar-broker/src/main/java/org/apache/pulsar/broker/service/ServerCnx.java
#	pulsar-common/src/main/java/org/apache/pulsar/common/api/proto/PulsarApi.java
#	pulsar-common/src/main/proto/PulsarApi.proto

+++ pulsar-broker/src/main/java/org/apache/pulsar/broker/loadbalance/impl/LinuxBrokerHostUsageImpl.java
@@ -55,7 +55,7 @@ public class LinuxBrokerHostUsageImpl implements BrokerHostUsage {
     private OperatingSystemMXBean systemBean;
     private SystemResourceUsage usage;
 
-    private final Optional<Integer> overrideBrokerNicSpeedGbps;
+    private final Optional<Double> overrideBrokerNicSpeedGbps;
 
     private static final Logger LOG = LoggerFactory.getLogger(LinuxBrokerHostUsageImpl.class);
